server/edge_anpr/requirements.txt
opencv-python==4.8.1.78
requests==2.31.0
python-dotenv==1.0.0
# Optional for better plate recognition:
# openalpr-python==3.1.0
# pytesseract==0.3.10
